Add some exceptions on add/remove vertex.

+++ src/Gliph/Graph/IncidenceGraph.php
@@ -11,6 +11,10 @@ class IncidenceGraph {
     }
 
     public function addVertex($vertex) {
+        if (!is_object($vertex)) {
+            throw new \OutOfBoundsException('Vertices must be objects; non-object provided.');
+        }
+
         if (!$this->hasVertex($vertex)) {
             $this->vertices[$vertex] = new \SplObjectStorage();
         }
@@ -24,6 +28,10 @@ class IncidenceGraph {
     }
 
     public function removeVertex($vertex) {
+        if (!$this->hasVertex($vertex)) {
+            throw new \OutOfRangeException('Vertex is not in the graph, it cannot be removed.', E_WARNING);
+        }
+
         foreach ($this->vertices[$vertex] as $adjacent) {
             $this->vertices[$adjacent]->detach($vertex);
         }
